apps/mobile/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]==3.2.9
python-dotenv==1.0.0
twilio==8.13.0
geopy==2.3.0
googlemaps==4.10.0
email-validator==2.1.0
redis==5.0.1
celery==5.3.4
